wmr89: New driver supporting the OS WMR89a
New driver for the WMR89a Weather Station from Oregon Scientific.
The actual sensor daata is captured an Arduino firmware, and then
relayed to WeeWX over a USB Serial connection. The data is encoded
in a way very similar to other OS devices, with minor differences
like byte order and checksum calculation.

This driver does *not* connect to the WMR89 basestation, so any stats
collected by the basestation are not collected.
WRM89a sensors supported
 * Oregon-THGR810 Temp-Hygro
 * Oregon-WGR800 Anemometer